--- fusedwake/gcl/python/__init__.py
@@ -1,1 +1,0 @@
-from .gcl import *
